Clearer structure, add copyrights
